Defer later profiles with overlapping local paths
Defer attempts to sync a profile which overlaps another syncing profile
until after the active command has finished. Introduces a new "sync due"
state for management of this.

Right now, "overlapping" means that one or more of a profile's local,
monitored paths is the same as that of another profile. This should
probably be extended to cover subdirectories.

Unison probably does this fairly robustly already via locking, but it still
makes sense to schedule setup and teardown of monitoring with knowledge of
profile "interdependencies".